Moves BMWlist to main folder
