+++ src/lib/Scenes/City/CityBMWList.tsx
@@ -3,7 +3,7 @@ import { PAGE_SIZE } from "lib/data/constants"
 import { isCloseToBottom } from "lib/utils/isCloseToBottom"
 import React from "react"
 import { createPaginationContainer, graphql, RelayPaginationProp } from "react-relay"
-import { EventList } from "./EventList"
+import { EventList } from "./components/EventList"
 
 interface Props {
   city: CityBMWList_city
@@ -49,7 +49,8 @@ class CityBMWList extends React.Component<Props, State> {
         key={name + "bmw"}
         cityName={name}
         bucket={shows.edges.map(e => e.node) as any}
-        type={"BMW Art Guide"}
+        header="BMW Art Guide"
+        type="BMW Art Guide"
         relay={relay}
         onScroll={isCloseToBottom(this.fetchData)}
         fetchingNextPage={fetchingNextPage}
